requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
boto3==1.35.54
botocore==1.35.54
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.6.0
contourpy==1.3.0
ctranslate2==4.4.0
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.7
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
faster-whisper==1.0.3
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gguf==0.10.0
google-api-core==2.23.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
interegular==0.3.3
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
lightning==2.4.0
lightning-cloud==0.5.70
lightning-utilities==0.11.8
lightning_sdk==0.1.30
linkify-it-py==2.0.3
litdata==0.2.19
litserve==0.2.3
llvmlite==0.43.0
lm-format-enforcer==0.10.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
mistral_common==1.4.4
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.20.0
open_clip_torch==2.29.0
openai==1.54.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
outlines==0.0.46
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pexpect==4.9.0
pillow==10.4.0
pip-autoremove==0.10.0
platformdirs==4.3.6
proglog==0.1.10
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.23.4
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.39.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.3
sentencepiece==0.2.0
simple-term-menu==1.6.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
textual==0.85.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.0.2
torch==2.4.0
torchmetrics==1.3.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.2
triton==3.0.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.1
vllm==0.6.3.post1
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
